+++ lib/server/controllers/QCFlagTypes.controller.js
@@ -15,7 +15,7 @@ const Joi = require('joi');
 const { DtoFactory } = require('../../domain/dtos/DtoFactory');
 const { dtoValidator } = require('../utilities/dtoValidator.js');
 const { updateExpressResponseFromNativeError } = require('../express/updateExpressResponseFromNativeError');
-const { qcFlagTypesService } = require('../services/qualityControlFlag/QcFlagTypesService');
+const { qcFlagTypesService } = require('../services/qualityControlFlag/QcFlagTypeService');
 
 // eslint-disable-next-line valid-jsdoc
 /**
